modified:   app/Http/Controllers/LandingController.php 	modified:   app/Http/Controllers/Profile/PhotoController.php 	new file:   app/Http/Controllers/SpotifyPlaylistController.php 	modified:   app/Models/Profile.php 	new file:   app/Models/SpotifyPlaylist.php 	new file:   database/migrations/2024_01_21_223321_create_spotify_playlists_table.php 	deleted:    public/images/user_photos/Messenger.png 	deleted:    public/images/user_photos/WhatsApp.png 	deleted:    public/images/user_photos/YouTube.png 	deleted:    public/images/user_photos/imo.png 	modified:   resources/views/landing.blade.php 	modified:   resources/views/pages/profile/photos/edit.blade.php 	modified:   routes/web.php

+++ app/Http/Controllers/Profile/PhotoController.php
@@ -88,15 +88,30 @@ class PhotoController extends Controller
         ]);
     }
 
-    public function destroy($id)
+    public function destroy($profile, $id)
     {
-        $photo = Photo::find($id);
-        dd($photo, 'destroy');
-        $photo->delete();
+        try {
+            $photo = Photo::find($id);
+            // dd($profile, $id, $photo, 'destroy');
+            // get photo path and delete it from public folder
+            $photoPath = public_path($photo->url);
+            // remove base url from photoPath
+            $photoPath = str_replace(config('app.url'), '', $photoPath);
+            if (file_exists($photoPath)) {
+                unlink($photoPath);
+                $photo->delete();
+            } else {
+                dd('file does not exist', 'destroy', $photoPath);
+            }
 
-        return redirect()->back()->with([
-            'message' => 'Photo deleted successfully',
-            'alert-type' => 'success'
-        ]);
+            return redirect()->back()->with([
+                'message' => 'Photo deleted successfully',
+                'alert-type' => 'success'
+            ]);
+        } catch (\Exception $e) {
+            return response()->json([
+                'error' => $e->getMessage()
+            ], 500);
+        }
     }
 }
